extract interaction handler into seperate action file
