+++ src/actions/index.ts
@@ -3,11 +3,13 @@ import * as registrationActions from 'src/actions/registration/'
 import * as navigationActions from 'src/actions/navigation/'
 import * as genericActions from 'src/actions/generic/'
 import * as ssoActions from 'src/actions/sso/'
+import * as interactionHandlerActions from 'src/actions/interactionHandler'
 
 export {
   accountActions,
   registrationActions,
   navigationActions,
   genericActions,
-  ssoActions
+  ssoActions,
+  interactionHandlerActions
 }
